Point A & B draw method is changed, so that only these points are redrawn and not the whole map. And init draw function is created to clear and bring it back to its original state. This avoids unnecessary condition checks & pixel drawing on the map. Point A & B are now filling up a whole square.

+++ src/main/java/be/uantwerpen/visualization/model/Cell.java
@@ -8,8 +8,8 @@ public class Cell {
     private int y;
     private String type;
     private String specific;
-    private float sur_x;
-    private float sur_y;
+    private int sur_x;
+    private int sur_y;
     private int spotID;
     /*
      * Type defenition
@@ -70,19 +70,19 @@ public class Cell {
         this.type = type;
     }
 
-    public float getSur_x() {
+    public int getSur_x() {
         return sur_x;
     }
 
-    public void setSur_x(float sur_x) {
+    public void setSur_x(int sur_x) {
         this.sur_x = sur_x;
     }
 
-    public float getSur_y() {
+    public int getSur_y() {
         return sur_y;
     }
 
-    public void setSur_y(float sur_y) {
+    public void setSur_y(int sur_y) {
         this.sur_y = sur_y;
     }
